add api calls to actions
more store complexity but reduces redeundant api calls from accidental caching and mutation actions.
also makes it easier to add websockets

- [x] api calls for spaces
- [x] api calls for cards
- [x] api calls for user
- [ ] api calls for connections
- [ ] api calls for connection types

- [x] refactor api options method to take options obj
- [x] DRY unique items id creation methods
- [x] add getSpace loader
- [x] update components to do dispatches -> api , instead of calling commits directly (which'll be reserved mainly for websockets that need to make a local update but not hit the api server)
- [x] currenspace: remove 'update' from method names to be more direct
- [ ] comment stub contributor key for allowing anon users to edit a remote space

- [x] QA all actions not signed in
- [ ] QA all signin actions
- [ ] QA all actions while signedin ,
- [ ] api queue adding should debounce and not 429 while editing cards

+++ src/api.js
@@ -1,7 +1,5 @@
 // https://www.notion.so/kinopio/API-docs
 
-// refactor options format
-
 import _ from 'lodash'
 
 import cache from '@/cache.js'
@@ -28,7 +26,7 @@ const api = {
     const body = currentUser
     body.email = email
     body.password = password
-    const options = this.options({ body, method: 'PATCH' })
+    const options = this.options({ body, method: 'POST' })
     try {
       const response = await fetch(`${host}/user/sign-up`, options)
       const normalizedResponse = await this.normalizeResponse(response)
@@ -42,7 +40,7 @@ const api = {
       email: email,
       password: password
     }
-    const options = this.options({ body, method: 'PATCH' })
+    const options = this.options({ body, method: 'POST' })
     try {
       const response = await fetch(`${host}/user/sign-in`, options)
       const normalizedResponse = await this.normalizeResponse(response)
@@ -53,7 +51,7 @@ const api = {
   },
   async resetPassword (email) {
     const body = { email }
-    const options = this.options({ body, method: 'PATCH' })
+    const options = this.options({ body, method: 'POST' })
     try {
       const response = await fetch(`${host}/user/reset-password`, options)
       const normalizedResponse = await this.normalizeResponse(response)
@@ -70,6 +68,8 @@ const api = {
     return utils.timeout(5000, fetch(`${host}/user`, options))
   },
   async removeUserPermanently () {
+    const userIsSignedIn = cache.user().apiKey
+    if (!userIsSignedIn) { return }
     try {
       const options = this.options({ method: 'DELETE' })
       await fetch(`${host}/user/permanent`, options)
@@ -91,11 +91,9 @@ const api = {
   // Card
 
   async createCard (body) {
-    const options = this.options({ body, method: 'PATCH' })
+    const options = this.options({ body, method: 'POST' })
     return utils.timeout(5000, fetch(`${host}/card`, options))
   },
-  // make this atomic w key/value instead?
-  // if needs whole card update => replaceCard?
   async updateCard (card) {
     const body = card
     const options = this.options({ body, method: 'PATCH' })
@@ -114,23 +112,56 @@ const api = {
     return utils.timeout(5000, fetch(`${host}/card/${cardId}/permanent`, options))
   },
 
+  // Connection
+
+  async updateConnection (connection) {
+    const body = connection
+    const options = this.options({ body, method: 'PATCH' })
+    return utils.timeout(5000, fetch(`${host}/connection/${connection.id}`, options))
+  },
+  // async removeConnections (body) {
+  //   console.log('🚨',body)
+  //   const options = this.options({ body, method: 'DELETE' })
+  //   return utils.timeout(5000, fetch(`${host}/connection/multiple`, options))
+  // },
+
+  // Connection Type
+
+  async updateConnectionType (connectionType) {
+    const body = connectionType
+    const options = this.options({ body, method: 'PATCH' })
+    return utils.timeout(5000, fetch(`${host}/connection-type/${connectionType.id}`, options))
+  },
+
   // Space
 
+  async updateSpace (space) {
+    const userIsSignedIn = cache.user().apiKey
+    // const userIsContributor = cache.space(space.id).contributorKey
+    if (!userIsSignedIn) { return } // (!userIsSignedIn && !userIsContributor)
+    console.log('🌙updateSpace', space)
+    const body = space
+    const options = this.options({ body, method: 'PATCH' })
+    return utils.timeout(5000, fetch(`${host}/card/${space.spaceId}`, options))
+  },
   async getSpace (spaceId) {
+    // TODO support getting other people's spaces later
+    const userIsSignedIn = cache.user().apiKey
+    if (!userIsSignedIn) { return }
+    console.log('🌙 Getting remote space', spaceId)
     const options = this.options({ method: 'GET' })
     const response = await utils.timeout(5000, fetch(`${host}/space/${spaceId}`, options))
     const normalizedResponse = await this.normalizeResponse(response)
     return normalizedResponse
   },
-
-  async saveSpace (body) {
-    const options = this.options({ body, method: 'PATCH' })
+  async createSpace (body) {
+    const options = this.options({ body, method: 'POST' })
     return utils.timeout(5000, fetch(`${host}/space`, options))
   },
-  async saveAllSpaces (apiKey) {
+  async saveSpaces (apiKey) {
     try {
       const body = cache.getAllSpaces()
-      const options = this.options({ body, apiKey, method: 'PATCH' })
+      const options = this.options({ body, apiKey, method: 'POST' })
       const response = await fetch(`${host}/space/multiple`, options)
       const normalizedResponse = await this.normalizeResponse(response)
       return normalizedResponse
@@ -158,14 +189,17 @@ const api = {
       body
     }
     queue.push(request)
-    console.log('in', request)
+    console.log('🎡 Add to request queue', request)
+    // consolidate
+    // if updateCard , ids are the same,
+
+    // other approach is to only add things to the api queue on stop interactions, when their done/blurred
     cache.saveQueue(queue)
-    this.processQueue()
+    _.debounce(_.wrap(this.processQueue()), 300)
   },
   request () {
     const queue = this.queue()
     const request = queue.shift()
-    console.log('out', request)
     cache.saveQueue(queue)
     return request
   },
@@ -176,6 +210,7 @@ const api = {
     return queue
   },
   async processQueue () {
+    console.log('this should be debounced')
     if (queueIsRunning) { return }
     if (!window.navigator.onLine) { return }
     let queue = this.queue()
@@ -186,11 +221,11 @@ const api = {
       try {
         request = this.request()
         await this.processRequest(request)
-        console.log('completed', request)
+        console.log('✅ completed', request)
       } catch (error) {
-        console.error(error)
         queue.push(request)
-        console.log('err re-in', request)
+        console.error(error)
+        console.log('🔁 Request error. Add back into queue to retry', request)
         cache.saveQueue(queue)
         queueIsRunning = false
         break
@@ -200,7 +235,7 @@ const api = {
     queueIsRunning = false
   },
   async processRequest (request) {
-    console.log('🚗', request.name, request.body)
+    console.log('🚎 Processing request', request.name, request.body)
     const response = await this[request.name](request.body)
     const normalizedResponse = await this.normalizeResponse(response)
     return normalizedResponse
@@ -217,9 +252,9 @@ const api = {
   // Request helpers
 
   options (options) {
-    console.log('🚎', options)
     const headers = new Headers({ 'Content-Type': 'application/json' })
-    const apiKey = options.apiKey || cache.user().apiKey
+    // const contributorKey = cache.space(options.spaceId).contributorKey
+    const apiKey = options.apiKey || cache.user().apiKey // || contributorKey
     if (apiKey) {
       headers.append('Authorization', apiKey)
     }
